Refresh unit UI and balance tweaks
Adds a new unit sprite sheet and updates queue/unit prefabs and scene layout for the refreshed HUD. Unit data assets now use the new icons and include upgrade settings, while stats and costs are tuned for Axeman, Cavalry, Scout, and Spearman units. Gold display is also rounded to an integer for cleaner UI output.

+++ Assets/_Project/Scripts/UI/UIManager.cs
@@ -171,7 +171,7 @@ namespace GridEmpire.UI
         private void RefreshGoldDisplay()
         {
             if (this == null || goldText == null || _localSpawner == null) return;
-            goldText.text = _localPlayer.Gold.ToString();
+            goldText.text = ((int)_localPlayer.Gold).ToString();
 
             if (goldPerTurnText != null)
                 goldPerTurnText.text = $"+{_localPlayer.GoldIncome:F1}";
